Move files from utils to lib.
