+++ lib/date-time.js
@@ -1,5 +1,5 @@
 var error = require('./error');
-var utils = require('../utils');
+var utils = require('./utils');
 
 var WEEK_STARTS = [
   undefined,
