+++ lib/engineering.js
@@ -1,9 +1,9 @@
 var error = require('./error');
 var jStat = require('jStat').jStat;
 var text = require('./text');
-var utils = require('../utils');
+var utils = require('./utils');
 var bessel = require('bessel');
-var _ = require('../utils/limited-lodash');
+var _ = require('./limited-lodash');
 
 function isValidBinaryNumber(number) {
   return (/^[01]{1,10}$/).test(number);
