+++ lib/expression-parser.js
