+++ lib/financial.js
@@ -1,7 +1,7 @@
 var error = require('./error');
 var dateTime = require('./date-time');
-var utils = require('../utils');
-var _ = require('../utils/limited-lodash');
+var utils = require('./utils');
+var _ = require('./limited-lodash');
 
 // TODO
 exports.ACCRINT = function () {
