+++ lib/information.js
@@ -1,4 +1,4 @@
-var utils = require('../utils');
+var utils = require('./utils');
 var error = require('./error');
 
 // TODO
