+++ lib/logical.js
@@ -1,7 +1,7 @@
 var error = require('./error');
-var utils = require('../utils');
+var utils = require('./utils');
 var information = require('./information');
-var _ = require('../utils/limited-lodash');
+var _ = require('./limited-lodash');
 
 exports.AND = function () {
   var args = _.flatten(arguments);
