+++ lib/lookup-reference.js
@@ -1,5 +1,5 @@
-var _ = require('../utils/limited-lodash');
-var utils = require('../utils');
+var _ = require('./limited-lodash');
+var utils = require('./utils');
 var error = require('./error');
 
 exports.CHOOSE = function (index_num) {
