+++ lib/math-trig.js
@@ -1,10 +1,10 @@
-var utils = require('../utils');
+var utils = require('./utils');
 var error = require('./error');
 var statistical = require('./statistical');
 var information = require('./information');
-var evalExpr = require('../utils/expression-parser');
-var matrix = require('../utils/matrix');
-var _ = require('../utils/limited-lodash');
+var evalExpr = require('./expression-parser');
+var matrix = require('./matrix');
+var _ = require('./limited-lodash');
 
 exports.ABS = function(number) {
   number = utils.parseNumber(number);
