+++ lib/matrix.js
