+++ lib/statistical.js
@@ -1,10 +1,10 @@
 var mathTrig = require('./math-trig');
 var text = require('./text');
 var jStat = require('jStat').jStat;
-var utils = require('../utils');
+var utils = require('./utils');
 var error = require('./error');
-var evalExpr = require('../utils/expression-parser');
-var _ = require('../utils/limited-lodash');
+var evalExpr = require('./expression-parser');
+var _ = require('./limited-lodash');
 
 var SQRT2PI = 2.5066282746310002;
