+++ lib/text.js
@@ -1,7 +1,7 @@
-var utils = require('../utils');
+var utils = require('./utils');
 var error = require('./error');
 var numeral = require('numeral');
-var _ = require('../utils/limited-lodash');
+var _ = require('./limited-lodash');
 
 //TODO
 exports.ASC = function () {
